requirements.txt
captcha==0.2.4
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-appconf==1.0.2
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-easy-select2==1.5.1
django-filter==1.1.0
django-imagekit==4.0.2
django-js-asset==1.1.0
django-paypal==0.5.0
django-recaptcha==1.4.0
django-select2==6.0.2
django-tinymce==2.7.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
idna==2.6
pilkit==2.0
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22

#channels==2.1.1
#channels_redis==2.2.1

google-api-python-client==1.7.4
oauth2client==4.1.2

django-social-share

#For linkpreview
beautifulsoup4


# pyjwt = for encoding and decoding JWT  Tokens (  for pusher Chatkit )
pyjwt